homework 3
