+++ php2homework4engine/controllers/Controller.php
@@ -3,12 +3,20 @@
 
 namespace app\controllers;
 
+use app\engine\Render;
+
 abstract class Controller
 {
     private $action;
     private $defaultAction = "index";
     private $layout = 'main';
     private $useLayouts = true;
+    private $renderer;
+
+    public function __construct()
+    {
+        $this->renderer = new Render();
+    }
 
     public function runAction($action = null) {
         $this->action = $action ?: $this->defaultAction;
@@ -31,11 +39,8 @@ abstract class Controller
         }
     }
 
-    public function renderTemplate($template, $params = []) {
-        ob_start();
-        extract($params);
-        $templatePath = TEMPLATES_DIR . $template . ".php";
-        include $templatePath;
-        return ob_get_clean();
+    public function renderTemplate($template, $params = [])
+    {
+        return $this->renderer->renderTemplate($template, $params);
     }
 }
