+++ php2homework4engine/engine/Render.php
@@ -1,0 +1,16 @@
+<?php
+
+
+namespace app\engine;
+
+
+class Render
+{
+    public function renderTemplate($template, $params = []) {
+        ob_start();
+        extract($params);
+        $templatePath = TEMPLATES_DIR . $template . ".php";
+        include $templatePath;
+        return ob_get_clean();
+    }
+}
